requirements.txt
flask
pillow
imagehash
gunicorn